Django-Middleware-0x03/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
drf-nested-routers>=0.93.5
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0

